requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.0
cssselect2==0.6.0
defusedxml==0.7.1
Django==4.0.4
django-cors-headers==3.11.0
django-pdf==0.0.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.1.0
future==0.18.2
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.2
html5lib==1.1
httplib2==0.20.4
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
lxml==4.9.0
MarkupSafe==2.1.1
oauthlib==3.2.0
oscrypto==1.3.0
Pillow==9.1.0
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.7
